todoproject/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-axes==8.3.1
django-mongodb-backend
sqlparse==0.5.5
